Add scenes folders and config
