+++ src/scenes/preloadScene.js
@@ -1,0 +1,11 @@
+import "phaser";
+
+export default class GameScene extends Phaser.Scene {
+  constructor() {
+    super("Game");
+  }
+
+  preload() {}
+
+  create() {}
+}
